requirements.txt
bumpversion
numpydoc
h5py
opencv
pandas
pillow
pip
psutil
pytorch
pyyaml>=5.1.0
pytest
matplotlib
numpy
recommonmark
rasterio
sphinx
slidingwindow
sphinx_rtd_theme
tqdm
twine
torchvision
xmltodict